auth_service/requirements.txt
fastapi==0.70.0
sqlalchemy==1.4.25
passlib[bcrypt]==1.7.4
python-jose==3.3.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
